Remove check before provisioning resources
We were checking that `(getAvailableVMs().size() > 0 || dags.size() >
0)`. This means that dags must be added to the WorkflowEngine before any
provisioning requests can happen. This means that provisioning requests
sent at t=0 may be silently ignored if they are carried out before the
`DAG_SUBMIT` event.

This lead to a difficult to understand bug in my PowerCappedProvisioner,
which uses provisioning requests for even the initial provisioning.

I'm not sure what the point of this check was (digging through the git
history gives no hints), but no tests fail if it is removed. This may
need more investigation...